CMake support, VS 2015 build fix, support for mpir instead of gmp, removed vc8 build, removed makefile, fixed stdint conflicts with modern compilers, changed README to reflect the changes
