+++ depends/include/README.txt
@@ -1,0 +1,2 @@
+gmp.h present in this directory was moved from the original bncsutil source code and is here for compatibility reasons.
+In order to use a newer version for modern compilers, see the main readme.
