+++ src/bncsutil/debug.h
@@ -95,10 +95,10 @@
 #ifndef __DEBUG_H__
 #define __DEBUG_H__
 
-#define bncsutil_debug_message(x)
-#define free_sys_err_msg(x)
-#define bncsutil_debug_message_a(x, y)
-#define sys_error_msg(x)
+//#define bncsutil_debug_message(x)
+//#define free_sys_err_msg(x)
+//#define bncsutil_debug_message_a(x, y)
+//#define sys_error_msg(x)
 
 
 #endif
